Need to fix texture mapping. Read further on Hyperbolic interpolation...

+++ RenderEngineProject/RenderEngineProject/Camera.h
@@ -4,7 +4,7 @@
 
 namespace RenderEngine
 {
-	extern const float PI;
+	extern const double PI;
 
 	class Camera
 	{
